refactor: 예외가 호출되는 위치를 고려 패키지 이동
