+++ backend/src/main/java/reviewme/review/service/module/ReviewDetailMapper.java
@@ -12,6 +12,7 @@ import reviewme.question.repository.QuestionRepository;
 import reviewme.review.domain.Review;
 import reviewme.review.domain.TextAnswer;
 import reviewme.review.domain.TextAnswers;
+import reviewme.review.domain.exception.OptionGroupNotFoundByQuestionIdException;
 import reviewme.review.service.dto.response.detail.OptionGroupAnswerResponse;
 import reviewme.review.service.dto.response.detail.OptionItemAnswerResponse;
 import reviewme.review.service.dto.response.detail.QuestionAnswerResponse;
@@ -19,7 +20,6 @@ import reviewme.review.service.dto.response.detail.ReviewDetailResponse;
 import reviewme.review.service.dto.response.detail.SectionAnswerResponse;
 import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.template.domain.Section;
-import reviewme.template.domain.exception.OptionGroupNotFoundByQuestionIdException;
 import reviewme.template.repository.SectionRepository;
 
 @Component
