+++ backend/src/main/java/reviewme/template/domain/exception/MissingOptionItemsInOptionGroupException.java
@@ -1,4 +1,4 @@
-package reviewme.question.domain.exception;
+package reviewme.template.domain.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import reviewme.global.exception.DataInconsistencyException;
