+++ backend/src/main/java/reviewme/template/service/module/TemplateMapper.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Component;
 import reviewme.question.domain.OptionGroup;
 import reviewme.question.domain.OptionItem;
 import reviewme.question.domain.Question;
-import reviewme.question.domain.exception.MissingOptionItemsInOptionGroupException;
 import reviewme.question.repository.OptionGroupRepository;
 import reviewme.question.repository.OptionItemRepository;
 import reviewme.question.repository.QuestionRepository;
@@ -15,6 +14,7 @@ import reviewme.template.domain.Section;
 import reviewme.template.domain.SectionQuestion;
 import reviewme.template.domain.Template;
 import reviewme.template.domain.TemplateSection;
+import reviewme.template.domain.exception.MissingOptionItemsInOptionGroupException;
 import reviewme.template.domain.exception.SectionInTemplateNotFoundException;
 import reviewme.template.domain.exception.TemplateNotFoundByReviewGroupException;
 import reviewme.template.repository.SectionRepository;
