+++ backend/src/test/java/reviewme/review/service/module/CheckBoxAnswerValidatorTest.java
@@ -15,11 +15,11 @@ import reviewme.question.repository.OptionGroupRepository;
 import reviewme.question.repository.OptionItemRepository;
 import reviewme.question.repository.QuestionRepository;
 import reviewme.review.domain.CheckboxAnswer;
+import reviewme.review.domain.exception.OptionGroupNotFoundByQuestionIdException;
 import reviewme.review.service.exception.CheckBoxAnswerIncludedNotProvidedOptionItemException;
 import reviewme.review.service.exception.SelectedOptionItemCountOutOfRangeException;
 import reviewme.review.service.exception.SubmittedQuestionNotFoundException;
 import reviewme.support.ServiceTest;
-import reviewme.template.domain.exception.OptionGroupNotFoundByQuestionIdException;
 
 @ServiceTest
 class CheckBoxAnswerValidatorTest {
