+++ backend/src/test/java/reviewme/template/service/TemplateMapperTest.java
@@ -16,7 +16,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import reviewme.question.domain.OptionGroup;
 import reviewme.question.domain.Question;
-import reviewme.question.domain.exception.MissingOptionItemsInOptionGroupException;
 import reviewme.question.repository.OptionGroupRepository;
 import reviewme.question.repository.OptionItemRepository;
 import reviewme.question.repository.QuestionRepository;
@@ -24,6 +23,7 @@ import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
 import reviewme.support.ServiceTest;
 import reviewme.template.domain.Section;
+import reviewme.template.domain.exception.MissingOptionItemsInOptionGroupException;
 import reviewme.template.domain.exception.SectionInTemplateNotFoundException;
 import reviewme.template.repository.SectionRepository;
 import reviewme.template.repository.TemplateRepository;
